refactor: domain, view 패키지 분리
