+++ docs/README.md
@@ -3,11 +3,16 @@
 - [x] 자동차 이름과 횟수 입력
 - [x] 자동차 전진
 - [x] 우승자 출력
-- [ ] 리팩토링
+- [x] 리팩토링
   - [x] 비즈니스 로직과 UI 로직 분리
   - [x] fianl 키워드 활용
   - [x] 접근 제한자 private 사용
-  - [ ] 단위 테스트 
+  - [x] 단위 테스트 
     - [x] 테스트 가능한 코드와 테스트하기 힘든 부분 분리
     - [x] 경계값을 기준으로 테스트
     - [x] 테스트 픽스처 생성
+  - [x] 패키지 분리
+    - 비즈니스 로직을 가지는 객체와 UI 관련 객체 패키지 분리
+  - [ ] 예외 처리
+    - [ ] 자동차 이름이 6자 이상인 경우 예외 처리
+    - [ ] 시도할 횟수가 숫자가 아닌 경우 예외 처리
