+++ src/main/java/racingcar/Application.java
@@ -1,9 +1,11 @@
 package racingcar;
 
-import java.util.ArrayList;
-import java.util.Collections;
+import racingcar.domain.Car;
+import racingcar.domain.Racing;
+import racingcar.view.InputView;
+import racingcar.view.ResultView;
+
 import java.util.List;
-import java.util.Scanner;
 
 public class Application {
